[libscc] Changed return type of treenode_numChildren from unsigned to size_t.

+++ src/libscc/TreeNode.cpp
@@ -1741,7 +1741,7 @@ YYLTYPE treenode_location(const TreeNode * node) {
     return toCxx(node)->location().toYYLTYPE();
 }
 
-unsigned treenode_numChildren(const TreeNode * node) {
+size_t treenode_numChildren(const TreeNode * node) {
     return toCxx(node)->children().size();
 }
